Fixed the embedded terminal, Fixed udp queue and added a way to view the udp queue in the graphical window.
Sometimes the embedded terminal wouldn't let us type into it, because of
some problem with the library. I couldn't pinpoint the exact reason for
this strange behaviour, but i fixed it by disabling autocomplete.

I also fixed the UDP queue, so now we can use it for getting callbacks
from the drone. We are still missing the timeout functionality but it
should be easy to implement i just haven't gotten around to doing it.

I also started creating a way to view the udp queue in the window, and i
added functionality for the udp server to print udp messages, that
aren't linked to the queue, to the embedded terminal. I don't know how
thread safe the ImTerm library is, but it seems to work. The initial
efforts to view telemetry data have started